Handle UTF8 input
We assume that the input file is UTF8, this means that the file needs to read in bytes and passed through to the JSON without re-encoding anything. encode_json will force a second encoding into utf8 by interpreting the bytes as individual codepoints which will result in the bytes `c3 a2 c2 86 c2 91` rather than the correct bytes `e2 86 91`.